// handwritten/firestore/dev/protos/firestore_v1_proto_api.d.ts
export namespace google {

    /** Namespace firestore. */
    namespace firestore {

        /** Namespace v1. */
        namespace v1 {
// ...
            interface IValue {

                /** Value nullValue */
                nullValue?: (google.protobuf.NullValue|null);

                /** Value booleanValue */
                booleanValue?: (boolean|null);

                /** Value integerValue */
                integerValue?: (number|string|null);

                /** Value doubleValue */
                doubleValue?: (number|null);

                /** Value timestampValue */
                timestampValue?: (google.protobuf.ITimestamp|null);

                /** Value stringValue */
                stringValue?: (string|null);

                /** Value bytesValue */
                bytesValue?: (Uint8Array|null);

                /** Value referenceValue */
                referenceValue?: (string|null);

                /** Value geoPointValue */
                geoPointValue?: (google.type.ILatLng|null);

                /** Value arrayValue */
                arrayValue?: (google.firestore.v1.IArrayValue|null);

                /** Value mapValue */
                mapValue?: (google.firestore.v1.IMapValue|null);

                /** Value fieldReferenceValue */
                fieldReferenceValue?: (string|null);

                /** Value functionValue */
                functionValue?: (google.firestore.v1.IFunction|null);

                /** Value pipelineValue */
                pipelineValue?: (google.firestore.v1.IPipeline|null);
            }
// ...
            interface IArrayValue {

                /** ArrayValue values */
                values?: (google.firestore.v1.IValue[]|null);
            }
// ...
            interface IMapValue {

                /** MapValue fields */
                fields?: ({ [k: string]: google.firestore.v1.IValue }|null);
            }
// ...
            interface IFunction {

                /** Function name */
                name?: (string|null);

                /** Function args */
                args?: (google.firestore.v1.IValue[]|null);

                /** Function options */
                options?: ({ [k: string]: google.firestore.v1.IValue }|null);
            }
// ...
            interface IPipeline {

                /** Pipeline stages */
                stages?: (google.firestore.v1.Pipeline.IStage[]|null);
            }
// ...
            namespace Pipeline {

                /** Properties of a Stage. */
                interface IStage {

                    /** Stage name */
                    name?: (string|null);

                    /** Stage args */
                    args?: (google.firestore.v1.IValue[]|null);

                    /** Stage options */
                    options?: ({ [k: string]: google.firestore.v1.IValue }|null);
                }
// ...
            }
// ...
            interface IExecutePipelineRequest {

                /** ExecutePipelineRequest database */
                database?: (string|null);

                /** ExecutePipelineRequest structuredPipeline */
                structuredPipeline?: (google.firestore.v1.IStructuredPipeline|null);

                /** ExecutePipelineRequest transaction */
                transaction?: (Uint8Array|null);

                /** ExecutePipelineRequest newTransaction */
                newTransaction?: (google.firestore.v1.ITransactionOptions|null);

                /** ExecutePipelineRequest readTime */
                readTime?: (google.protobuf.ITimestamp|null);
            }
// ...
            interface IStructuredPipeline {

                /** StructuredPipeline pipeline */
                pipeline?: (google.firestore.v1.IPipeline|null);

                /** StructuredPipeline options */
                options?: ({ [k: string]: google.firestore.v1.IValue }|null);
            }
// ...
        }
    }
// ...
    namespace protobuf {
// ...
        type NullValue =
            "NULL_VALUE";
// ...
        interface ITimestamp {

            /** Timestamp seconds */
            seconds?: (number|string|null);

            /** Timestamp nanos */
            nanos?: (number|null);
        }
// ...
    }

    /** Namespace type. */
    namespace type {

        /** Properties of a LatLng. */
        interface ILatLng {

            /** LatLng latitude */
            latitude?: (number|null);

            /** LatLng longitude */
            longitude?: (number|null);
        }
// ...
    }
// ...
}